Fix BottomBar clipping and reposition ProgressBar
This commit addresses two separate UI issues:

1.  **BottomBar clipping on mobile:** The BottomBar was being clipped on mobile devices due to the browser's dynamic UI. This was fixed by adding a script to `app/page.tsx` that sets the `--app-height` CSS variable to the window's inner height on load and resize. This ensures the application viewport is always correctly sized.

2.  **Incorrect ProgressBar position:** The video progress bar was located at the very bottom of the screen. The desired behavior was for it to be attached to the top of the BottomBar. This was resolved by refactoring the components:
    - The progress bar logic and JSX were moved from `VideoPlayer.tsx` to `BottomBar.tsx`.
    - The `videoRef` is now created in `Slide.tsx` and passed down to both child components, allowing `BottomBar` to control and react to the video's state.
    - The progress bar is now positioned absolutely within `BottomBar` to achieve the desired visual overlap, as seen in the `tingtong.txt` prototype.

These changes align with the provided diagnosis and patch, resulting in a more robust and visually correct user interface on all devices.
