+++ app/page.tsx
@@ -32,6 +32,16 @@ export default function Home() {
     fetchSlides();
   }, []);
 
+  // Fix 1: Dynamically set app height for mobile browsers
+  useEffect(() => {
+    const setAppHeight = () => {
+      document.documentElement.style.setProperty('--app-height', `${window.innerHeight}px`);
+    };
+    setAppHeight();
+    window.addEventListener('resize', setAppHeight);
+    return () => window.removeEventListener('resize', setAppHeight);
+  }, []);
+
   useEffect(() => {
     const observer = new IntersectionObserver(
       (entries) => {
